+++ core/src/main/java/io/bitsquare/payment/CryptoCurrencyAccountContractData.java
@@ -31,7 +31,7 @@ public final class CryptoCurrencyAccountContractData extends PaymentAccountContr
     @Nullable
     private String paymentId;
 
-    public CryptoCurrencyAccountContractData(String paymentMethod, String id, int maxTradePeriod) {
+    public CryptoCurrencyAccountContractData(String paymentMethod, String id, long maxTradePeriod) {
         super(paymentMethod, id, maxTradePeriod);
     }
